Moved python files to their own directory. All tests passing; gui working aside from IBT issues.
