+++ bbpow/decoder.py
@@ -6,11 +6,11 @@ from hashlib import sha256
 
 from basicblockchains_ecc.elliptic_curve import secp256k1
 
-from block import Block
-from formatter import Formatter
-from headers import Header
-from transactions import Transaction, MiningTransaction
-from utxo import UTXO_INPUT, UTXO_OUTPUT
+from .block import Block
+from .formatter import Formatter
+from .headers import Header
+from .transactions import Transaction, MiningTransaction
+from .utxo import UTXO_INPUT, UTXO_OUTPUT
 
 
 class Decoder:
